+++ src/draw/DefaultDrawer.js
@@ -1,0 +1,13 @@
+/**
+* Class that draws a component in a graphical representation.
+*/
+class DefaultDrawer {
+  /**
+  * Returns an svg from a component and an svg template.
+  */
+  draw() {
+    return null;
+  }
+}
+
+export default DefaultDrawer;