+++ src/index.js
@@ -4,6 +4,10 @@ import ComponentAttributeDefinition from './models/ComponentAttributeDefinition'
 import ComponentDrawOption from './models/ComponentDrawOption';
 import ComponentAttribute from './models/ComponentAttribute';
 import ComponentLink from './models/ComponentLink';
+import DefaultDrawer from './draw/DefaultDrawer';
+import DefaultMetadata from './metadata/DefaultMetadata';
+import DefaultParser from './parser/DefaultParser';
+import DefaultRender from './render/DefaultRender';
 
 export {
   Component,
@@ -12,4 +16,8 @@ export {
   ComponentDrawOption,
   ComponentAttribute,
   ComponentLink,
+  DefaultDrawer,
+  DefaultMetadata,
+  DefaultParser,
+  DefaultRender,
 };